use uint32 instead of uint8 for pin type

+++ src/ArduinoHal.cpp
@@ -18,31 +18,31 @@ void ArduinoHal::term() {
   }
 }
 
-void inline ArduinoHal::pinMode(uint8_t pin, uint8_t mode) {
+void inline ArduinoHal::pinMode(uint32_t pin, uint32_t mode) {
   if (pin == RADIOLIB_NC) {
     return;
   }
   ::pinMode(pin, RADIOLIB_ARDUINOHAL_PIN_MODE_CAST mode);
 }
-void inline ArduinoHal::digitalWrite(uint8_t pin, uint8_t value) {
+void inline ArduinoHal::digitalWrite(uint32_t pin, uint32_t value) {
   if (pin == RADIOLIB_NC) {
     return;
   }
   ::digitalWrite(pin, RADIOLIB_ARDUINOHAL_PIN_STATUS_CAST value);
 }
-uint8_t inline ArduinoHal::digitalRead(uint8_t pin) {
+uint32_t inline ArduinoHal::digitalRead(uint32_t pin) {
   if (pin == RADIOLIB_NC) {
     return 0;
   }
   return ::digitalRead(pin);
 }
-void inline ArduinoHal::attachInterrupt(uint8_t interruptNum, void (*interruptCb)(void), uint8_t mode) {
+void inline ArduinoHal::attachInterrupt(uint32_t interruptNum, void (*interruptCb)(void), uint32_t mode) {
   if (interruptNum == RADIOLIB_NC) {
     return;
   }
   ::attachInterrupt(interruptNum, interruptCb,  RADIOLIB_ARDUINOHAL_INTERRUPT_MODE_CAST mode);
 }
-void inline ArduinoHal::detachInterrupt(uint8_t interruptNum) {
+void inline ArduinoHal::detachInterrupt(uint32_t interruptNum) {
   if (interruptNum == RADIOLIB_NC) {
     return;
   }
@@ -60,7 +60,7 @@ unsigned long inline ArduinoHal::millis() {
 unsigned long inline ArduinoHal::micros() {
   return ::micros();
 }
-long inline ArduinoHal::pulseIn(uint8_t pin, uint8_t state, unsigned long timeout) {
+long inline ArduinoHal::pulseIn(uint32_t pin, uint32_t state, unsigned long timeout) {
   if (pin == RADIOLIB_NC) {
     return 0;
   }
@@ -81,7 +81,7 @@ void inline ArduinoHal::spiEndTransaction() {
 void inline ArduinoHal::spiEnd() {
   _spi->end();
 }
-void inline ArduinoHal::tone(uint8_t pin, unsigned int frequency, unsigned long duration) {
+void inline ArduinoHal::tone(uint32_t pin, unsigned int frequency, unsigned long duration) {
   #if !defined(RADIOLIB_TONE_UNSUPPORTED)
   if (pin == RADIOLIB_NC) {
     return;
@@ -107,7 +107,7 @@ void inline ArduinoHal::tone(uint8_t pin, unsigned int frequency, unsigned long 
   pwmPin->write(0.5);
   #endif
 }
-void inline ArduinoHal::noTone(uint8_t pin) {
+void inline ArduinoHal::noTone(uint32_t pin) {
   #if !defined(RADIOLIB_TONE_UNSUPPORTED) and defined(ARDUINO_ARCH_STM32)
   if (pin == RADIOLIB_NC) {
     return;
@@ -140,7 +140,7 @@ void inline ArduinoHal::yield() {
   ::yield();
   #endif
 }
-uint8_t inline ArduinoHal::pinToInterrupt(uint8_t pin) {
+uint32_t inline ArduinoHal::pinToInterrupt(uint32_t pin) {
   return digitalPinToInterrupt(pin);
 }
 #endif
